Pressure Sensitivity
Makes changes to the default pressure sensitivity modifier value to half of what it originally was, now when using the `Apply Pressure` modifier it results in the corresponding buttons to register a "tap" whereas with the default value it would register as a "hold" no matter what.

Made changes to the Enable XInput Code instead of just omitting it completely.